+++ backend/domain_types/src/connector_types.rs
@@ -93,6 +93,7 @@ pub enum ConnectorEnum {
     Billwerk,
     Hipay,
     Trustpayments,
+    Redsys,
     Globalpay,
     Nuvei,
     Iatapay,
@@ -101,6 +102,7 @@ pub enum ConnectorEnum {
     Paybox,
     Barclaycard,
     Nexixpay,
+    Mollie,
     Airwallex,
     Tsys,
     Bankofamerica,
@@ -175,7 +177,9 @@ impl ForeignTryFrom<grpc_api_types::payments::Connector> for ConnectorEnum {
             grpc_api_types::payments::Connector::Nmi => Ok(Self::Nmi),
             grpc_api_types::payments::Connector::Shift4 => Ok(Self::Shift4),
             grpc_api_types::payments::Connector::Barclaycard => Ok(Self::Barclaycard),
+            grpc_api_types::payments::Connector::Redsys => Ok(Self::Redsys),
             grpc_api_types::payments::Connector::Nexixpay => Ok(Self::Nexixpay),
+            grpc_api_types::payments::Connector::Mollie => Ok(Self::Mollie),
             grpc_api_types::payments::Connector::Airwallex => Ok(Self::Airwallex),
             grpc_api_types::payments::Connector::Tsys => Ok(Self::Tsys),
             grpc_api_types::payments::Connector::Bankofamerica => Ok(Self::Bankofamerica),
@@ -1087,6 +1091,9 @@ pub struct PaymentsAuthorizeData<T: PaymentMethodDataTypes> {
     pub payment_channel: Option<PaymentChannel>,
     pub enable_partial_authorization: Option<bool>,
     pub locale: Option<String>,
+    pub redirect_response: Option<ContinueRedirectionResponse>,
+    pub threeds_method_comp_ind: Option<ThreeDsCompletionIndicator>,
+    pub continue_redirection_url: Option<Url>,
     pub tokenization: Option<common_enums::Tokenization>,
 }
 
@@ -1340,12 +1347,14 @@ pub enum PaymentsResponseData {
         status_code: u16,
     },
     PreAuthenticateResponse {
+        authentication_data: Option<router_request_types::AuthenticationData>,
         /// For Device Data Collection
         redirection_data: Option<Box<RedirectForm>>,
         connector_response_reference_id: Option<String>,
         status_code: u16,
     },
     AuthenticateResponse {
+        resource_id: Option<ResponseId>,
         /// For friction flow
         redirection_data: Option<Box<RedirectForm>>,
         /// For frictionles flow
@@ -1439,6 +1448,22 @@ pub struct PaymentsPreAuthenticateData<T: PaymentMethodDataTypes> {
     pub browser_info: Option<BrowserInformation>,
     pub enrolled_for_3ds: bool,
     pub redirect_response: Option<ContinueRedirectionResponse>,
+    pub capture_method: Option<common_enums::CaptureMethod>,
+}
+
+impl<T: PaymentMethodDataTypes> PaymentsPreAuthenticateData<T> {
+    pub fn is_auto_capture(&self) -> Result<bool, Error> {
+        match self.capture_method {
+            Some(common_enums::CaptureMethod::Automatic)
+            | None
+            | Some(common_enums::CaptureMethod::SequentialAutomatic) => Ok(true),
+            Some(common_enums::CaptureMethod::Manual) => Ok(false),
+            Some(common_enums::CaptureMethod::ManualMultiple)
+            | Some(common_enums::CaptureMethod::Scheduled) => {
+                Err(ConnectorError::CaptureMethodNotSupported.into())
+            }
+        }
+    }
 }
 
 #[derive(Debug, Clone)]
@@ -1453,6 +1478,35 @@ pub struct PaymentsAuthenticateData<T: PaymentMethodDataTypes> {
     pub browser_info: Option<BrowserInformation>,
     pub enrolled_for_3ds: bool,
     pub redirect_response: Option<ContinueRedirectionResponse>,
+    pub capture_method: Option<common_enums::CaptureMethod>,
+    pub authentication_data: Option<router_request_types::AuthenticationData>,
+}
+
+impl<T: PaymentMethodDataTypes> PaymentsAuthenticateData<T> {
+    pub fn is_auto_capture(&self) -> Result<bool, Error> {
+        match self.capture_method {
+            Some(common_enums::CaptureMethod::Automatic)
+            | None
+            | Some(common_enums::CaptureMethod::SequentialAutomatic) => Ok(true),
+            Some(common_enums::CaptureMethod::Manual) => Ok(false),
+            Some(common_enums::CaptureMethod::ManualMultiple)
+            | Some(common_enums::CaptureMethod::Scheduled) => {
+                Err(ConnectorError::CaptureMethodNotSupported.into())
+            }
+        }
+    }
+
+    pub fn get_browser_info(&self) -> Result<BrowserInformation, Error> {
+        self.browser_info
+            .clone()
+            .ok_or_else(missing_field_err("browser_info"))
+    }
+
+    pub fn get_continue_redirection_url(&self) -> Result<Url, Error> {
+        self.continue_redirection_url
+            .clone()
+            .ok_or_else(missing_field_err("continue_redirection_url"))
+    }
 }
 
 #[derive(Debug, Clone)]
@@ -1478,6 +1532,20 @@ pub struct PaymentsSdkSessionTokenData {
     pub shipping_cost: Option<MinorUnit>,
     /// The specific payment method type for which the session token is being generated
     pub payment_method_type: Option<PaymentMethodType>,
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize)]
+/// Indicates if 3DS method data was successfully completed or not
+pub enum ThreeDsCompletionIndicator {
+    /// 3DS method successfully completed
+    #[serde(rename = "Y")]
+    Success,
+    /// 3DS method was not successful
+    #[serde(rename = "N")]
+    Failure,
+    /// 3DS method URL was unavailable
+    #[serde(rename = "U")]
+    NotAvailable,
 }
 
 #[derive(Debug, Clone)]